+++ test/lua_tests/singleton_access.lua
@@ -1,1 +1,2 @@
 local engine = assert(Engine, "Getting singleton Engine failed")
+local autoload = assert(GlobalNode, "Getting autoload node failed")